pyproject.toml
[tool.poetry]
name = "github-pr-notification"
version = "0.1.0"
description = ""
authors = ["sinofseven <em.s.00001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "1.55"
boto3-stubs = {extras = ["dynamodb", "ssm", "cloudformation"], version = "1.24.42"}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipython = "^8.4.0"
boto3 = "^1.24.42"
prompt-toolkit = "^3.0.30"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
